agents/email_responder/gmail-consent-app/pyproject.toml
[project]
name = "gmail-receiver"
version = "0.1.0"
description = "Cloud Run push receiver for Gmail Pub/Sub notifications"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.5",
    "uvicorn[standard]==0.32.0",
    "google-auth==2.36.0",
    "google-auth-oauthlib==1.2.1",
    "google-cloud-pubsub==2.23.0",
    "google-cloud-secret-manager==2.21.1",
    "httpx==0.27.2",
    "pydantic==2.9.2",
    "python-dotenv==1.0.1",
    "google-api-python-client==2.150.0"
]

[tool.uv]
package = true

[tool.uv.sources]
gmail-receiver = {path = "src"}


